Make it build using cl

+++ src/Utils.cpp
@@ -250,7 +250,7 @@ std::string SplitBucket::getBakedPattern(Index value) const {
   std::tie(prefix, suffix) = getInternalPrefixAndSuffix(pattern);
   const auto padding = pattern.size() - prefix.size() - suffix.size();
   char buffer[10]; // 4,294,967,295 is 10 characters long maximum.
-  StringView view(buffer, padding == 1 ? 1 + std::log10(value >=1 ? value : 1) : padding);
+  StringView view(buffer, padding == 1 ? 1 + static_cast<size_t>(std::log10(value >=1 ? value : 1)) : padding);
   bake(value, view);
   return concat(prefix, view, suffix);
 }
@@ -269,7 +269,7 @@ char getStep(Indices sortedIndices) {
       }
     }
   }
-  assert(minimum_step <= std::numeric_limits<char>::max());
+  assert(minimum_step <= static_cast<size_t>(std::numeric_limits<char>::max()));
   return minimum_step_set ? static_cast<char>(minimum_step) : -1;
 }
